rename upload action and reducer to reports for clarity
