+++ tdrs-frontend/src/reducers/index.js
@@ -4,7 +4,7 @@ import alert from './alert'
 import auth from './auth'
 import stts from './sttList'
 import requestAccess from './requestAccess'
-import upload from './upload'
+import reports from './reports'
 
 /**
  * Combines all store reducers
@@ -17,7 +17,7 @@ const rootReducer = (history) =>
     auth,
     stts,
     requestAccess,
-    upload,
+    reports,
   })
 
 export default rootReducer
